perf + seo: fix LCP, server-seed SEO pages, add DB indexes + caching
Session 2 addresses LCP (4.4s → target <2.5s), indexing issues on Google
Search Console ("Discovered - currently not indexed"), API timeouts on
/api/leagues/[slug]/schedule (3.8% timeout / 29.3% CPU throttle), and
thin per-page SEO metadata.

P1: Homepage LCP
- Hero Image renders unconditionally (was gated on client state, preventing
  priority preload emission). Added sizes="100vw", quality=75, blur placeholder.
- Video deferred until window.load + requestIdleCallback, preload="none",
  fade-in when canplaythrough fires. Image stays as the LCP paint.
- next.config.js: enabled AVIF, 30-day minimumCacheTTL, tuned device sizes.
- app/layout.js: metadataBase, OG/Twitter images, expanded SportsOrganization
  JSON-LD with full address, geo, areaServed.

P2: SSR for SEO
- New lib/server/leagues.js — shared unstable_cache-wrapped data layer
  (getActiveLeagues, getLeagueSchedule, getLeagueStandings, getLatestResults)
  with tag-based invalidation per league slug.
- app/standings, /schedule, /results: server pages now fetch initial league
  data and pass it as initialData prop. Client components skip first useEffect
  fetch when seeded. Initial HTML now contains real team names/scores for
  Googlebot.
- generateStaticParams on /leagues/[slug], /standings/[slug], /schedule/[slug],
  /results/[slug] — pre-renders 3 league slugs at build time.
- All multi-league pages swapped from dynamic='force-dynamic' to revalidate=300.
- Homepage is now a Server Component that server-fetches latest results;
  hero/interactive bits moved to components/home/HomePageClient.js.
- Added sr-only SEO paragraphs to /standings, /schedule, /results with
  keyword-rich copy.

P3: API timeouts + DB
- prisma/schema.prisma: added 16 @@index declarations on unindexed FKs
  and hot filter columns (Season.leagueId+status+seasonNumber,
  Match.weekId+status, 3 team-FK indexes on Match, TierPlacement.weekId/
  tierId, Week.seasonId+status, League.isActive, PlayerStat.matchId,
  Player.teamId, tournament FKs).
- prisma/migrations/20260416_add_performance_indexes/migration.sql:
  hand-crafted idempotent CREATE INDEX IF NOT EXISTS + ANALYZE. Apply to
  prod via `npx prisma db execute --file ...`.
- lib/prisma.js: reads POSTGRES_PRISMA_URL first, falls back to DATABASE_URL.
  Singleton now enabled in all envs (was dev-only), fixing connection pool
  exhaustion on warm Vercel instances.
- API routes delegate to cached functions — unstable_cache TTL 300s
  (schedule/standings), 60s (latest results).
- Admin mutations (scores, tier-movement, tier-placements) call
  revalidateLeagueByWeek() to bust caches on score entry.

P5: Per-page metadata
- All routes now have keyword-rich titles (Mississauga + Volleyball), unique
  150–160 char descriptions, OG + Twitter cards with hero image, canonical.
- /rules gets FAQPage JSON-LD (5 Q&As) for rich snippet eligibility.
- /champions and /register use layout.js for metadata (pages are 'use client').

Verified
- npm run build: 21 static + 9 SSG league pages generated
- npm run test: 24/24 passing
- npm run test:e2e: 43/43 passing

Follow-up for user
- Apply migration SQL against prod DB
- Confirm POSTGRES_PRISMA_URL is set in Vercel env
- Re-submit indexed pages in Google Search Console
- (Optional) compress source images offline
