+++ lib/prisma.js
@@ -1,15 +1,25 @@
 import { PrismaClient } from "@prisma/client";
 import { PrismaPg } from "@prisma/adapter-pg";
 
+// Cache on globalThis in BOTH dev and prod. On Vercel, each warm serverless
+// instance re-uses this module between invocations — without the singleton,
+// every request creates a new pg pool and eventually exhausts the per-role
+// connection limit on Prisma Postgres.
 const globalForPrisma = globalThis;
 
 function createClient() {
+  // Prefer the pooled connection URL provided by Vercel Postgres integration
+  // (POSTGRES_PRISMA_URL routes through PgBouncer), fall back to DATABASE_URL.
+  let connectionString =
+    process.env.POSTGRES_PRISMA_URL ||
+    process.env.DATABASE_URL ||
+    '';
+
   // Fix SSL warning: replace sslmode=require with sslmode=verify-full
-  let connectionString = process.env.DATABASE_URL || '';
   connectionString = connectionString.replace('sslmode=require', 'sslmode=verify-full');
 
   // Cap pool size to stay under Prisma Postgres' per-role connection limit.
-  // Default (cpus * 2 + 1) is too high for hosted dev databases with low caps.
+  // With PgBouncer multiplexing, one connection per warm instance is enough.
   const adapter = new PrismaPg({
     connectionString,
     max: 3,
@@ -18,13 +28,7 @@ function createClient() {
   return new PrismaClient({ adapter });
 }
 
-// Reuse a single Prisma client + adapter across hot reloads in development.
-// Creating a new adapter on every reload leaks connection pools until the
-// database role's connection limit is exhausted.
 const prisma = globalForPrisma.prisma ?? createClient();
-
-if (process.env.NODE_ENV !== "production") {
-  globalForPrisma.prisma = prisma;
-}
+globalForPrisma.prisma = prisma;
 
 export default prisma;
