+++ next.config.js
@@ -3,6 +3,15 @@ const nextConfig = {
   typescript: {
     ignoreBuildErrors: true,
   },
+  images: {
+    // Prefer AVIF (smaller) then WebP then original format.
+    formats: ['image/avif', 'image/webp'],
+    // Cache optimized images for 30 days at the edge.
+    minimumCacheTTL: 2592000,
+    // Device breakpoints tuned for mobile-first volleyball audience.
+    deviceSizes: [360, 640, 750, 828, 1080, 1200, 1440, 1920, 2048],
+    imageSizes: [16, 32, 48, 64, 96, 128, 256, 384],
+  },
   async headers() {
     return [
       {
